rewrite server to ts
