+++ package.json
@@ -16,10 +16,12 @@
     "gui": "hermione gui"
   },
   "devDependencies": {
-    "@babel/core": "^7.6.2",
+    "@babel/core": "^7.6.4",
     "@babel/plugin-proposal-class-properties": "^7.5.5",
-    "@babel/preset-env": "^7.6.2",
+    "@babel/preset-env": "^7.6.3",
     "@babel/preset-react": "^7.0.0",
+    "@types/express": "^4.17.1",
+    "@types/node": "^12.7.12",
     "babel-loader": "^8.0.6",
     "chai": "^4.2.0",
     "copy-webpack-plugin": "^5.0.4",
@@ -36,16 +38,18 @@
     "sass-loader": "^8.0.0",
     "style-loader": "^1.0.0",
     "svg-react-loader": "^0.4.6",
+    "ts-loader": "^6.2.0",
+    "typescript": "^3.6.4",
     "url-loader": "^2.1.0",
-    "webpack": "^4.41.0",
+    "webpack": "^4.41.1",
     "webpack-cli": "^3.3.9"
   },
   "dependencies": {
     "axios": "^0.19.0",
     "classnames": "^2.2.6",
     "express": "^4.17.1",
-    "react": "^16.10.1",
-    "react-dom": "^16.10.1",
+    "react": "^16.10.2",
+    "react-dom": "^16.10.2",
     "react-redux": "^7.1.1",
     "react-router-dom": "^5.1.2",
     "redux": "^4.0.4",
