+++ src/client/store/actions.js
@@ -1,4 +1,7 @@
-import { SET_DIRECTORY_CONTENT, SET_DIRECTORY_PATH, SET_REPOSITORIES, SET_CURRENT_REPOSITORY, SET_FILE_DATA, CLEAR_FILE_DATA } from "./actionTypes";
+import { 
+    SET_DIRECTORY_CONTENT, SET_DIRECTORY_PATH, SET_REPOSITORIES, SET_CURRENT_REPOSITORY,
+    SET_FILE_DATA, CLEAR_FILE_DATA
+} from "./actionTypes";
 
 
 export const setRepositories = (repos) => ({
