+++ webpack-server.config.js
@@ -7,20 +7,28 @@ const DEBUG = process.env.NODE_ENV !== 'production';
 const SRC_PATH = path.resolve(__dirname, 'src');
 const BUILD_PATH = path.resolve(__dirname, 'build');
 const PUBLIC_PATH = path.resolve(SRC_PATH, 'public');
-const SERVER_PATH = path.resolve(BUILD_PATH,  'server');
+const SERVER_PATH = path.resolve(BUILD_PATH, 'server');
 
 let config = {
-    mode: DEBUG ? 'development': 'production',
-    entry:  path.resolve(SRC_PATH, 'server/index.js'),
+    mode: DEBUG ? 'development' : 'production',
+    entry: path.resolve(SRC_PATH, 'server/index.ts'),
     target: 'node',
     context: SRC_PATH,
-    output:  {
+    output: {
         path: SERVER_PATH,
         filename: 'index.js',
     },
     resolve: {
-        extensions: ['.js'],
+        extensions: ['.ts', '.js'],
         modules: [path.resolve('node_modules'), SRC_PATH]
+    },
+    module: {
+        rules: [
+            {
+                test: /\.tsx?$/,
+                loader: "ts-loader"
+            }
+        ]
     },
     plugins: [
         new CopyPlugin([{
